+++ examples/gate_nand.py
@@ -1,24 +1,19 @@
-import numpy
+import random
 import nufhe
-from reikna.cluda import any_api
-
-thr = any_api().Thread.create(interactive=True)
-
-rng = numpy.random.RandomState()
-secret_key, cloud_key = nufhe.make_key_pair(thr, rng)
 
 size = 32
+bits1 = [random.choice([False, True]) for i in range(size)]
+bits2 = [random.choice([False, True]) for i in range(size)]
+reference = [not (b1 and b2) for b1, b2 in zip(bits1, bits2)]
 
-bits1 = rng.randint(0, 2, size=size).astype(numpy.bool)
-bits2 = rng.randint(0, 2, size=size).astype(numpy.bool)
+ctx = nufhe.Context()
+secret_key, cloud_key = ctx.make_key_pair()
 
-ciphertext1 = nufhe.encrypt(thr, rng, secret_key, bits1)
-ciphertext2 = nufhe.encrypt(thr, rng, secret_key, bits2)
+ciphertext1 = ctx.encrypt(secret_key, bits1)
+ciphertext2 = ctx.encrypt(secret_key, bits2)
 
-reference = ~(bits1 * bits2)
+vm = ctx.make_virtual_machine(cloud_key)
+result = vm.gate_nand(ciphertext1, ciphertext2)
+result_bits = ctx.decrypt(secret_key, result)
 
-result = nufhe.empty_ciphertext(thr, cloud_key.params, ciphertext1.shape)
-nufhe.gate_nand(thr, cloud_key, result, ciphertext1, ciphertext2)
-
-result_bits = nufhe.decrypt(thr, secret_key, result)
-assert (result_bits == reference).all()
+assert all(result_bits == reference)