+++ doc/source/history.rst
@@ -8,6 +8,8 @@ Version history
 
 * **CHANGED:** a ``PerformanceParameters`` object needs to be specialized for the device used (by calling its ``for_device()`` method) before passing it to gates.
 
+* ADDED: a high-level API hiding the Reikna details and removing some boilerplate.
+
 * FIXED: a bug in ``tlwe_noiseless_trivial()`` occasionally leading to memory corruption.
 
 * FIXED: a bug where ``PerformanceParameters`` and ``PerformanceParametersForDevice`` objects did not have a correct equality implementation, leading to unnecessary re-compilation of kernels.